User authentication for backend

+++ backend/Controllers/runController.js
@@ -1,12 +1,12 @@
 const asyncHandler = require('express-async-handler')
-
 const Run = require('../models/runModel')
+const User = require('../models/userModel')
 
 //@desc   Get Runs
 //@route  GET /api/runsData
 //@access Private
 const getRuns = asyncHandler(async (req, res) => {
-  const runs = await Run.find()
+  const runs = await Run.find({ user: req.user.id })
 
   res.status(200).json(runs)
 })
@@ -19,6 +19,7 @@ const setRun = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Please fill in all fields')
   }
+
   const run = await Run.create({
     date: req.body.date,
     runTime: req.body.runTime,
@@ -27,6 +28,7 @@ const setRun = asyncHandler(async (req, res) => {
     avgHeartRate: req.body.avgHeartRate,
     activeCalories: req.body.activeCalories,
     totalCalories: req.body.activeCalories,
+    user: req.user.id
   })
 
   res.status(200).json(run)
@@ -43,6 +45,19 @@ const updateRun = asyncHandler(async (req, res) => {
     throw new Error('Run not found')
   }
 
+  const user = await User.findById(req.user.id)
+  //Check for user
+  if(!user){
+    res.status(401)
+    throw new Error('User Not Found')
+  }
+
+  //Make sure the logged in user matches the goal user
+  if(run.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User Not Authorized')
+  }
+
   const updatedRun = await Run.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
   res.status(200).json(updatedRun)
@@ -57,6 +72,19 @@ const deleteRun = asyncHandler(async (req, res) => {
   if (!run) {
     res.status(400)
     throw new Error('Run not found')
+  }
+
+  const user = await User.findById(req.user.id)
+  //Check for user
+  if(!user){
+    res.status(401)
+    throw new Error('User Not Found')
+  }
+
+  //Make sure the logged in user matches the goal user
+  if(run.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User Not Authorized')
   }
 
   const deletedRun = await Run.findByIdAndDelete(req.params.id)
